cleaning (gcc warnings)

+++ src/benchmarks/multi.cpp
@@ -38,7 +38,7 @@ struct zdt1 {
     Eigen::VectorXd res(2);
     double f1 = x(0);
     double g = 1.0;
-    for (size_t i = 1; i < x.size(); ++i)
+    for (int i = 1; i < x.size(); ++i)
       g += 9.0 / (x.size() - 1) * x(i);
     double h = 1.0f - sqrtf(f1 / g);
     double f2 = g * h;
@@ -54,7 +54,7 @@ struct zdt2 {
     Eigen::VectorXd res(2);
     double f1 = x(0);
     double g = 1.0;
-    for (size_t i = 1; i < x.size(); ++i)
+    for (int i = 1; i < x.size(); ++i)
       g += 9.0 / (x.size() - 1) * x(i);
     double h = 1.0f - pow((f1 / g), 2.0);
     double f2 = g * h;
@@ -70,7 +70,7 @@ struct zdt3 {
     Eigen::VectorXd res(2);
     double f1 = x(0);
     double g = 1.0;
-    for (size_t i = 1; i < x.size(); ++i)
+    for (int i = 1; i < x.size(); ++i)
       g += 9.0 / (x.size() - 1) * x(i);
     double h = 1.0f - sqrtf(f1 / g) - f1 / g * sin(10 * M_PI * f1);
     double f2 = g * h;
